5_rest_api/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-filter==24.2
sqlparse==0.4.4
typing_extensions==4.11.0
